LayoutTests/imported/w3c/web-platform-tests/svg/extensibility/foreignObject/getboundingclientrect-expected.txt


FAIL foreignObject: Element.prototype.getBoundingClientRect() assert_equals: top expected 24 but got 18
